Adobe Lightroom Midi Controller - V0.1a
This is a temporary sketch to understand needed code to build a simple Midi Controller which later will be used as an Adobe Ligthroom Controller.
last modified on 17 Jun 2017
by midijon

All buttons will be implemented using Shift Registers (595 type) or Muxes (4067 type).

I decided not to use bounce.h for now, and I have absolutely no idea why (maybe because I need to understand a bit better the way it works before)

This sketch includes:
• 1 Teensy 3.6
• 16 Rotary Encoders KY-040 type
• 5 LEDs
• 3 Buttons

Will be added later :
• 1 additional Rotary Encoder
• 21 additional Buttons
• possibly 8 RGB leds
• possibly a 2x16 LCD screen